Remove Newtonsoft.Json from Settings (#393)
* Remove Newtonsoft.Json from Settings

* Fix codeql

* Revert "Fix codeql"

This reverts commit 04d83dd35318f6e56230695c53f49c1826c1b6f8.

* Enable windows targeting to build on linux

https://learn.microsoft.com/en-us/dotnet/core/tools/sdk-errors/netsdk1100

* Use system.text.json

* Use json converters that don't require runtime dynamic code

* Add specialized converters

* Use init

* Format

* Use correct enum member type

* Only support stuff that's valid json

* Add static converters

* Use switch

* Remove unused build config

+++ src/HolzShots.Core/Input/Selection/AreaSelector.cs
@@ -47,7 +47,7 @@ public partial class AreaSelector : AnimatedForm, IAreaSelector
         windowEnumerationTask.CancelAfter(5000);
         WindowFinder.GetCurrentWindowRectanglesAsync(Handle, allowEntireScreen, windowEnumerationTask.Token).ContinueWith(t => SetAvailableWindows(t.Result));
 
-        _dimmingOverlayBrush = Device.CreateSolidColorBrush(new D2DColor(settingsContext.AreaSelectorDimmingOpacity, OverlayColor)) ?? throw new D2DInteropException($"Could not create `{nameof(D2DBitmap)}`");
+        _dimmingOverlayBrush = Device.CreateSolidColorBrush(new D2DColor((float)settingsContext.AreaSelectorDimmingOpacity, OverlayColor)) ?? throw new D2DInteropException($"Could not create `{nameof(D2DBitmap)}`");
         _image = Device.CreateBitmapFromGDIBitmap(image) ?? throw new D2DInteropException($"Could not create `{nameof(D2DBitmap)}`");
         _imageBounds = new Rectangle(0, 0, image.Width, image.Height);
         _dimmedImage = CreateDimmedImage(image.Width, image.Height);
